Use `keyWindow` instead of `delegate`
This problem is not new and was already solved here #396. I am using
exaclty the same setup and unfortunatelly the sharing controller is
hidden underneath the inappbrowser controller and the `getTopMostViewController`
method fails to identify the correct `presentingViewController`.

By changing the `UIViewController` to be obtained from the `delegate` to
the `keyWindow` everything seems to be working as expected.