Update plugin.xml
To solve this error;
Build failed with the following error: There was an issue generating the app. Elements of type "receiver" with defined intent filters must declare a value for "android:exported" in all AndroidManifest.xml files, both in the app and Gradle dependencies. Check your plugin configurations and try again.